bug-deriv-engine/deriv_engine/requirements.txt
websockets==12.0
supabase==2.0.1
python-dotenv==1.0.0
numpy==1.26.4
scipy==1.12.0
